+++ .storybook/preview.tsx
@@ -45,9 +45,9 @@ import 'components/common/DatabaseEditor/widgets/menu/subMenuOption.scss';
 import 'components/common/DatabaseEditor/widgets/menuWrapper.scss';
 import 'components/common/DatabaseEditor/widgets/propertyMenu.scss';
 import 'components/common/DatabaseEditor/widgets/checkbox.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-vars.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-indent.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-list.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-indent.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-list.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-vars.scss';
 import 'theme/databases/databases.button.scss';
 import 'theme/databases/databases.main.scss';
 import 'react-resizable/css/styles.css';
