--- components/common/CharmEditor/components/listItemNew/czi-indent.scss
@@ -1,59 +1,0 @@
-/* Indent */
-
-// invlude div in selector so that it overrides the margin set in czi-list.scss with :not() selectors
-div.ProseMirror {
-  --list-item-indent: 24px;
-}
-
-/* list item indents */
-
-div.ProseMirror ol[data-indent='1'],
-div.ProseMirror ul[data-indent='1'] {
-  margin-left: var(--list-item-indent);
-}
-
-div.ProseMirror ol[data-indent='2'],
-div.ProseMirror ul[data-indent='2'] {
-  margin-left: calc(2 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='3'],
-div.ProseMirror ul[data-indent='3'] {
-  margin-left: calc(3 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='4'],
-div.ProseMirror ul[data-indent='4'] {
-  margin-left: calc(4 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='5'],
-div.ProseMirror ul[data-indent='5'] {
-  margin-left: calc(5 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='6'],
-div.ProseMirror ul[data-indent='6'] {
-  margin-left: calc(6 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='7'],
-div.ProseMirror ul[data-indent='7'] {
-  margin-left: calc(7 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='8'],
-div.ProseMirror ul[data-indent='8'] {
-  margin-left: calc(8 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='9'],
-div.ProseMirror ul[data-indent='9'] {
-  margin-left: calc(9 * var(--list-item-indent));
-}
-
-div.ProseMirror ol[data-indent='10'],
-div.ProseMirror ul[data-indent='10'] {
-  margin-left: calc(10 * var(--list-item-indent));
-}
-
