--- components/common/CharmEditor/components/listItemNew/markNames.ts
@@ -1,3 +1,0 @@
-export const MARK_FONT_SIZE = 'mark-font-size';
-export const MARK_TEXT_COLOR = 'mark-text-color';
-export const MARK_TEXT_SELECTION = 'mark-text-selection';
