--- components/common/CharmEditor/components/listItemNew/nodeNames.ts
@@ -1,8 +1,0 @@
-export const BLOCKQUOTE = 'blockquote';
-export const BULLET_LIST = 'bullet_list';
-export const HEADING = 'heading';
-export const LIST_ITEM = 'list_item';
-export const ORDERED_LIST = 'ordered_list';
-export const PARAGRAPH = 'paragraph';
-
-export const TEXT = 'text';
