+++ pages/_app.tsx
@@ -90,9 +90,9 @@ import 'components/common/DatabaseEditor/widgets/menu/separatorOption.scss';
 import 'components/common/DatabaseEditor/widgets/menu/subMenuOption.scss';
 import 'components/common/DatabaseEditor/widgets/menuWrapper.scss';
 import 'components/common/DatabaseEditor/widgets/propertyMenu.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-indent.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-list.scss';
-import 'components/common/CharmEditor/components/listItemNew/czi-vars.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-indent.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-list.scss';
+import 'packages/charmeditor/src/extensions/listItem/czi-vars.scss';
 import 'prosemirror-menu/style/menu.css';
 import 'react-resizable/css/styles.css';
 import 'theme/@bangle.dev/styles.scss';
